+++ src/Prairie.hs
@@ -5,10 +5,20 @@ module Prairie
     ( module Prairie.Class
     , module Prairie.Update
     , module Prairie.Diff
+    , module Prairie.Fold
+    , module Prairie.Traverse
+    , module Prairie.Zip
     , module Prairie.TH
+    , module Prairie.Semigroup
+    , module Prairie.Monoid
     ) where
 
 import Prairie.Class
-import Prairie.Update
 import Prairie.Diff
+import Prairie.Fold
+import Prairie.Monoid
+import Prairie.Semigroup
 import Prairie.TH
+import Prairie.Traverse
+import Prairie.Update
+import Prairie.Zip