+++ src/Prairie/Diff.hs
@@ -2,12 +2,12 @@
 --
 -- @since 0.0.1.0
 module Prairie.Diff
-  ( module Prairie.Diff
-  , module Prairie.Update
-  ) where
+    ( module Prairie.Diff
+    , module Prairie.Update
+    ) where
 
-import Prairie.Update
 import Prairie.Class
+import Prairie.Update
 
 -- | Given two 'Record's, this function produces a list of 'Update's that
 -- can be performed on the first record such that it will equal the second.
@@ -22,20 +22,19 @@ import Prairie.Class
 --
 -- @since 0.0.1.0
 diffRecord
-  :: (Record rec, FieldDict Eq rec)
-  => rec
-  -- ^ The old record.
-  -> rec
-  -- ^ The new record.
-  -> [Update rec]
-diffRecord old new = foldMap go allFields
+    :: (Record rec, FieldDict Eq rec)
+    => rec
+    -- ^ The old record.
+    -> rec
+    -- ^ The new record.
+    -> [Update rec]
+diffRecord old new = foldMap go allFieldsList
   where
     go (SomeField f) =
-      withFieldDict @Eq f $
-        let
-          newVal = getRecordField f new
-         in
-          if getRecordField f old /= newVal
-            then [SetField f newVal]
-            else []
-
+        withFieldDict @Eq f $
+            let
+                newVal = getRecordField f new
+             in
+                if getRecordField f old /= newVal
+                    then [SetField f newVal]
+                    else []